+++ src/main/scala/ship/computer/internals/InstructionControl.scala
@@ -4,11 +4,11 @@ import ship.computer.IntcodeProgram
 
 
 trait InstructionControl { self: IntcodeProgram =>
-  private lazy val instructionPointerValue: Int = address(instructionPointer)
+  private lazy val instructionPointerValue: Long = address(instructionPointer)
 
-  def opcode: Int = Math.abs(instructionPointerValue % 100)
-  def parameterMode(parameterNumber: Int): Int = Math.abs(instructionPointerValue / Math.pow(10, parameterNumber + 1).toInt) % 10
+  def opcode: Long = Math.abs(instructionPointerValue % 100)
+  def parameterMode(parameterNumber: Long): Long = Math.abs(instructionPointerValue / Math.pow(10, parameterNumber + 1).toInt) % 10
 
-  def moveInstructionPointer(increment: Int): IntcodeProgram = setInstructionPointer(instructionPointer + increment)
-  def setInstructionPointer(address: Int): IntcodeProgram = copy(instructionPointer = address)
+  def moveInstructionPointer(increment: Long): IntcodeProgram = setInstructionPointer(instructionPointer + increment)
+  def setInstructionPointer(address: Long): IntcodeProgram = copy(instructionPointer = address)
 }
