+++ src/main/scala/ship/thrusters/amplifiers/Amplifier.scala
@@ -4,10 +4,10 @@ import ship.computer.IntcodeProgram
 
 
 case class Amplifier(phaseSetting: Int, var amplificationProgram: IntcodeProgram) {
-  def lastOutput: Int = amplificationProgram.configuration.output.trackedOutput.last
+  def lastOutput: Long = amplificationProgram.configuration.output.trackedOutput.last
   def isComplete: Boolean = amplificationProgram.state.isComplete
 
-  def run(input: Int = 0): Amplifier = {
+  def run(input: Long = 0): Amplifier = {
     amplificationProgram = amplificationProgram.continue(input)
     this
   }
