+++ src/main/scala/ship/thrusters/amplifiers/AmplifierResult.scala
@@ -1,7 +1,7 @@
 package ship.thrusters.amplifiers
 
 
-case class AmplifierResult(phaseSettings: Seq[Int] = Vector.empty, result: Int = 0) {
+case class AmplifierResult(phaseSettings: Seq[Int] = Vector.empty, result: Long = 0) {
   def max(other: AmplifierResult): AmplifierResult =
     if(result > other.result) this else other
